Examples/hackathonProjects/DentricNewer/hack/requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
datasets==2.15.0
perforatedai==3.0.7
wandb==0.15.8
scikit-learn==1.3.1
matplotlib==3.8.0
seaborn==0.13.0
pandas==2.0.3
numpy==1.24.3
Pillow==10.0.0
tqdm==4.66.1